feat: update `caller` Rename caller to client. Move `requests` inside functions.
